requirements.txt
Django==4.2
selenium==4.9.0
gunicorn==20.1.0
Fabric3==1.14.post1